+++ src/components/auth/AuthForm.tsx
@@ -7,12 +7,12 @@ import { Loader2 } from "lucide-react";
 
 import { Button } from "@/components/ui/button";
 import { Form } from "@/components/ui/form";
-
-import { authFormSchema } from "@/lib/validations";
-import FillIcon from "../icons/FillIcons";
-import { Separator } from "../ui/separator";
 import CustomInput from "../shared/CustomInput";
+import FillIcon from "../icons/FillIcons";
+
+import { Separator } from "../ui/separator";
 import { useAuthForm } from "@/hooks/auth/useAuthForm";
+import { authFormSchema } from "@/lib/validations/index";
 
 const AuthForm = ({ type }: { type: "signup" | "signin" }) => {
   const formSchema = authFormSchema(type);
