+++ src/lib/validations/index.ts
@@ -1,0 +1,2 @@
+export * from "./authSchema";
+export * from "./checkoutSchema";
